Refactor aria-pressed
