+++ theme/base/javascripts/selectors.js
@@ -91,5 +91,5 @@ export const doneButtonClass = 'previousSelection__done';
 export const searchingClass = 'wayf__idpList--searching';
 export const noResultSectionSelector = '.wayf__noResults';
 export const idpListSelector = '.wayf__idpList';
-export const ariaPressedLabelSelector = 'label[aria-pressed]';
+export const ariaPressedCheckboxSelector = 'input[aria-pressed]';
 export const topId = 'top';
