+++ theme/base/templates/modules/Authentication/View/Proxy/Partials/WAYF/preselection.html.twig
@@ -6,7 +6,7 @@
     <input type="checkbox" name="previousSelectionToggle" id="previousSelectionToggle" class="visually-hidden previousSelection__toggle" tabindex="-1" role="button" aria-pressed="false" />
     {% spaceless %}
     <div class="ie11__label">
-        <label for="previousSelectionToggle" tabindex="0" class="previousSelection__toggleLabel" aria-pressed="false">
+        <label for="previousSelectionToggle" tabindex="0" class="previousSelection__toggleLabel">
             <span class="previousSelection__edit">{{ 'edit'|trans }}</span>
             <span class="previousSelection__done">{{ 'done'|trans }}</span>
         </label>
